+++ test/spmd/checkpoint/test_dt_planner.py
@@ -16,10 +16,9 @@ from spmd.testing.checkpoint_utils import with_temp_dir
 from spmd.testing.common_utils import (
     DistTensorTestBase,
     with_comms,
+    skip_unless_torch_gpu,
 )
 from torch.testing._internal.common_utils import run_tests
-
-from spmd.testing.devices import skip_unless_torch_gpu
 
 
 class MyModule(torch.nn.Module):
