+++ test/spmd/tensor/test_dtensor_ops.py
@@ -27,8 +27,9 @@ from spmd.testing.common_utils import (
     DistTensorTestBase,
     TEST_SKIPS,
     DTensorConverter,
+    DEVICE_TYPE,
+    NUM_DEVICES,
 )
-from spmd.testing.devices import DEVICE_TYPE, NUM_DEVICES
 
 # rewrite common size variables to sth can be sharded evenly
 # we can enable uneven shards later, but need to adjust more on
